traced_perf: re-enable linux build + add x86 regs handling
We need the kernel UAPI headers for the perf regs enums, I'm sourcing
them from buildtools/bionic, as with the in-tree builds.

I've previously promised to look into explicit reassignments instead of
memcpy/loop for arm/arm64, but ended up keeping them, as there's no
really sensible enum to index the "user" regs by (I'm not exactly sure
whether MachineArm64.h is the conceptually right enum either, but that's
what's used in practice, and all of these layouts match anyway).

Bug: 144281346

+++ src/profiling/perf/event_reader.cc
@@ -56,7 +56,8 @@ static int perf_event_open(perf_event_attr* attr,
 // cpu-scoped?
 base::ScopedFile PerfEventOpen(const EventConfig& event_cfg) {
   base::ScopedFile perf_fd{
-      perf_event_open(event_cfg.perf_attr(), /*pid=*/-1, event_cfg.target_cpu(),
+      perf_event_open(event_cfg.perf_attr(), /*pid=*/-1,
+                      static_cast<int>(event_cfg.target_cpu()),
                       /*group_fd=*/-1, PERF_FLAG_FD_CLOEXEC)};
   return perf_fd;
 }
